pyproject.toml
[project]
name = "domuscura"
version = "0.1.0"
description = "Self-hosted home maintenance tracker"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "django>=5.1,<6.0",
    "django-htmx>=1.21",
    "httpx>=0.27",
    "gunicorn>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "pytest-playwright>=0.6",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "domuscura.settings"
pythonpath = ["."]
addopts = "-m 'not e2e'"
markers = [
    "e2e: end-to-end browser tests (require Playwright)",
]
